Moves Icon component to shared
