+++ src/components/shared/Icon/Icon.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import './Icon.scss';
 
-const Icon = ({ icon, cssClasses = "" }) => (
+const Icon = ({ icon, cssClasses = '' }) => (
   <svg className={`Icon ${cssClasses}`} viewBox={icon.viewBox}>
     <path d={icon.path} />
   </svg>
