+++ src/components/shared/Icon/Icon.scss
